+++ .pathplanner/settings.json
@@ -9,8 +9,8 @@
   ],
   "autoFolders": [],
   "defaultMaxVel": 4.0,
-  "defaultMaxAccel": 6.0,
+  "defaultMaxAccel": 3.0,
   "defaultMaxAngVel": 540.0,
   "defaultMaxAngAccel": 720.0,
-  "maxModuleSpeed": 7.0
+  "maxModuleSpeed": 5.4
 }
