+++ src/main/java/frc/robot/RobotContainer.java
@@ -78,7 +78,7 @@ public class RobotContainer
       new CommandXboxController(OIConstants.kCoDriverControllerPort);
   
   private final SwerveSubsystem drivebase = new SwerveSubsystem(new File(Filesystem.getDeployDirectory(),
-                                                                       "swerve/neo-test"));
+                                                                       "swerve/neo"));
