+++ src/main/java/frc/robot/subsystems/ElevatorSubsystem.java
@@ -27,7 +27,6 @@ public class ElevatorSubsystem extends SubsystemBase {
 
     public ElevatorSubsystem(){ 
         m_elevator.setInverted(true);
-        m_elevator_helper.setInverted(true);
         elevatorEncoder.setPosition(0);
         m_PidController.setFeedbackDevice(elevatorEncoder);
         m_elevator.setIdleMode(IdleMode.kBrake);
@@ -39,7 +38,7 @@ public class ElevatorSubsystem extends SubsystemBase {
         m_PidController.setIZone(kIz);
         m_PidController.setFF(kFF);
         m_PidController.setOutputRange(kMinOutput, kMaxOutput);
-        m_elevator_helper.follow(m_elevator, false);
+        m_elevator_helper.follow(m_elevator, true);
 
     }
