+++ src/main/java/frc/robot/subsystems/swervedrive/SwerveSubsystem.java
@@ -4,6 +4,7 @@
 
 package frc.robot.subsystems.swervedrive;
 
+import java.util.ArrayList;
 import com.pathplanner.lib.auto.AutoBuilder;
 import com.pathplanner.lib.commands.PathPlannerAuto;
 import com.pathplanner.lib.path.PathConstraints;
@@ -45,6 +46,8 @@ public class SwerveSubsystem extends SubsystemBase
    * Swerve drive object.
    */
   private final SwerveDrive swerveDrive;
+
+  private final GyroHistory history;
   /**
    * Maximum speed of the robot in meters per second, used to limit acceleration.
    */
@@ -86,6 +89,7 @@ public class SwerveSubsystem extends SubsystemBase
     //swerveDrive.setCosineCompensator(!SwerveDriveTelemetry.isSimulation); // Disables cosine compensation for simulations since it causes discrepancies not seen in real life.
     swerveDrive.setCosineCompensator(false);
     setupPathPlanner();
+    history = new GyroHistory(swerveDrive);
   }
 
   /**
@@ -97,6 +101,7 @@ public class SwerveSubsystem extends SubsystemBase
   public SwerveSubsystem(SwerveDriveConfiguration driveCfg, SwerveControllerConfiguration controllerCfg)
   {
     swerveDrive = new SwerveDrive(driveCfg, controllerCfg, maximumSpeed);
+    history = new GyroHistory(swerveDrive);
   }
 
   /**
@@ -305,7 +310,7 @@ public class SwerveSubsystem extends SubsystemBase
   @Override
   public void periodic()
   {
-    recordGyroData();
+    history.recordGyroData();
     addVisionReading();
   }
 
@@ -502,75 +507,19 @@ public class SwerveSubsystem extends SubsystemBase
 
   public void addVisionReading()
   {
-    //if(LimelightHelpers.getTV("limelight-april")) {
-    if(true) {
+    if(LimelightHelpers.getTV("limelight-april")) {
+    //if(true) {
       Pose2d limelightPose = LimelightHelpers.getBotPose2d_wpiBlue("limelight-april");
-      //double llLatency = LimelightHelpers.getLatency_Capture("limelight-april")/1000 + LimelightHelpers.getLatency_Pipeline("limelight-april")/1000;
-      double llLatency = 0.25;
+      double llLatency = LimelightHelpers.getLatency_Capture("limelight-april")/1000 + LimelightHelpers.getLatency_Pipeline("limelight-april")/1000;
+      //double llLatency = 0.25;
 
-      System.out.println("LL Latency: " + llLatency);
+     
 
     swerveDrive.addVisionMeasurement(
       new Pose2d(limelightPose.getX(), limelightPose.getY(), 
-                 fetchInterpolatedGyroData(llLatency)),
+                 history.fetchInterpolatedGyroData(llLatency)),
                  Timer.getFPGATimestamp()- llLatency
                  );
     }
   }
-
-  private double[][] gyroData = new double[50][2];
-
-  public void recordGyroData() {
-    for(int x = gyroData.length-2; x>=0; x--) {
-      gyroData[x+1][0] = gyroData[x][0];
-      gyroData[x+1][1] = gyroData[x][1];
-    }
-    //gyroData[0][0] = swerveDrive.field.getRobotPose().getRotation().getDegrees();
-    gyroData[0][0] = (Timer.getFPGATimestamp()*325)%360;
-    gyroData[0][1] = Timer.getFPGATimestamp();
-  }
-
-  public Rotation2d fetchInterpolatedGyroData(double llLatency) {
-
-    double wantedTimestamp = Timer.getFPGATimestamp() - llLatency;
-    int bestID = -1;
-    double bestDiff = Double.MAX_VALUE; // Initialize with maximum value
-
-     // Find the index of the closest timestamp
-    for(int x = 0; x < gyroData.length; x++) {
-        double diff = Math.abs(gyroData[x][1] - wantedTimestamp);
-        if(diff < bestDiff) {
-            bestDiff = diff;
-            bestID = x;
-        }
-    }
-    System.out.println("the closest real data point is array id " + bestID);
-    System.out.println("searching for: " + wantedTimestamp);
-
-    //this will return the raw value if its on the edge, which should never happen, but saftey
-    if (bestID == 0 || bestID == gyroData.length - 1) {
-      return Rotation2d.fromDegrees(gyroData[bestID][0]);
-    }
-
-    double prevTimestamp = gyroData[bestID - 1][1];
-    double nextTimestamp = gyroData[bestID + 1][1];
-    double prevValue = gyroData[bestID - 1][0];
-    double nextValue = gyroData[bestID + 1][0];
-    System.out.println("values nearby: " + prevValue + ", " + gyroData[bestID][0] + ", " + nextValue);
-    System.out.println("times nearby: " + prevTimestamp + ", " + gyroData[bestID][1] + ", " + nextTimestamp);
-
-    if ((Math.abs(prevValue-nextValue))>100) {
-      return Rotation2d.fromDegrees(gyroData[bestID][0]);
-    }
-
-    double interpolatedValue = interpolate(prevTimestamp, prevValue, nextTimestamp, nextValue, wantedTimestamp);
-    System.out.println("interpreted value is " + interpolatedValue);
-    System.out.println("---");
-    return Rotation2d.fromDegrees(interpolatedValue);
-  }
-
-  private double interpolate(double x0, double y0, double x1, double y1, double x) {
-    return y0 + (y1 - y0) * ((x - x0) / (x1 - x0));
-  }
-
 };
